+++ components/ImageSwiper.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import React, { useState, useEffect } from "react";
-import { Navigation, Pagination } from "swiper/modules";
+import { Navigation } from "swiper/modules";
 import { Swiper, SwiperSlide } from "swiper/react";
 import "swiper/css";
 import "swiper/css/navigation";
@@ -9,17 +9,7 @@ import "@/styles/swiper.css";
 
 const ImageSwiper = ({ urls }: { urls: string[] }) => {
   const [showModal, setShowModal] = useState(false);
-  const [clickImageUrl, setClickImageUrl] = useState("");
   const [currentIndex, setCurrentIndex] = useState(1);
-
-  const modalOpen = (imageUrl: string) => {
-    setShowModal(true);
-    setClickImageUrl(imageUrl);
-  };
-
-  const modalClose = () => {
-    setShowModal(false);
-  };
 
   useEffect(() => {
     setCurrentIndex(1);
@@ -48,7 +38,7 @@ const ImageSwiper = ({ urls }: { urls: string[] }) => {
                 fill
                 sizes="(width: 435px, height: 220px)"
                 onClick={() => {
-                  modalOpen(url.toString());
+                  setShowModal(true);
                 }}
                 className="cursor-pointer object-cover"
               />
@@ -64,7 +54,7 @@ const ImageSwiper = ({ urls }: { urls: string[] }) => {
       {showModal && (
         <div
           className="fixed z-50 left-0 top-0 w-full h-full overflow-auto bg-black bg-opacity-40 flex items-center justify-center"
-          onClick={modalClose}
+          onClick={() => setShowModal(false)}
         >
           <div className="relative p-1 w-3/4 max-w-lg mx-auto h-auto rounded-md" onClick={e => e.stopPropagation()}>
             <Swiper
